data-cleaning-image/requirements.txt
boto3
pyspark  
python-dotenv